dedup test
